+++ lib/components/drawer.dart
@@ -2,6 +2,7 @@
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:mylastwords/Background/tracker.dart';
 import 'package:mylastwords/Screens/AboutScreen/about_screen.dart';
 import 'package:mylastwords/Screens/GalleryScreen/gallery_screen.dart';
 import 'package:mylastwords/Screens/Login/login_screen.dart';
@@ -9,8 +10,9 @@ import 'package:mylastwords/Screens/NoteScreen/note_screen.dart';
 import 'package:mylastwords/Screens/ProfileScreen/profile_screen.dart';
 import 'package:mylastwords/Services/user_service.dart';
 import 'package:mylastwords/components/rounded_button.dart';
-import 'package:mylastwords/components/subscription.dart';
+import 'package:mylastwords/components/subscribed_screen.dart';
 import 'package:mylastwords/components/toastmessage.dart';
+import 'package:mylastwords/components/unsubscribed_screen.dart';
 import 'package:mylastwords/constants.dart';
 import 'package:mylastwords/models/api_response.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -30,9 +32,10 @@ class _DrawerScreenState extends State<DrawerScreen> {
   String userImage = '';
   String userEmail = '';
   bool? isLoggedIn;
+  String subscription = '';
   
   @override
-  void initState() {
+  void initState() {    
     loadData();
     super.initState();
   }
@@ -44,6 +47,7 @@ class _DrawerScreenState extends State<DrawerScreen> {
       userImage = (prefs.getString('userImage') ?? '');  
       isLoggedIn = (prefs.getBool('isLoggedIn') ?? false);    
       userEmail = (prefs.getString('email') ?? '');
+      subscription = (prefs.getString('subcription')??'');
     });    
   }
    @override
@@ -80,9 +84,16 @@ class _DrawerScreenState extends State<DrawerScreen> {
             Icons.subscriptions,
           ),
           title: const Text('Subscription'),
-          onTap: () {
-            if(isLoggedIn==true){
-              Navigator.of(context).pushAndRemoveUntil(MaterialPageRoute(builder: (context) => SubscriptionScreen()),(route) => false);
+          onTap: () {          
+            if(isLoggedIn==true){         
+              if(subscription=="free")
+              {
+                Navigator.of(context).pushAndRemoveUntil(MaterialPageRoute(builder: (context) => SubscriptionScreen()),(route) => false);
+              }
+              else if(subscription=="subscribed"){
+                Navigator.of(context).pushAndRemoveUntil(MaterialPageRoute(builder: (context) => UnsubscribedScreen()),(route) => false);
+              }
+              
             }            
           },
         ),
